// src/app/models/employee.model.ts
export class Employee {
    id?: number;
    firstName?: string;
    lastName?: string;
    email?: string;
    phone?: number;
    salary?: number;
    birthdate?: Date;
    gender?: boolean;
    imageURL?: string;
    companyName?: string;
    jobExperience?: string;
    length: any;
    attachment: any
}